Se modifica los permisos de la app

+++ src/components/FileDownloaderScreen.jsx
@@ -1,56 +1,62 @@
 import React, {useState} from 'react';
-import {View, Text, Button} from 'react-native';
+import {View, Text, Button, Alert} from 'react-native';
 import {Picker} from '@react-native-picker/picker';
 import XLSX from 'xlsx';
 import RNFS from 'react-native-fs';
 import {handleGetData} from '../assets/questionnaireFunctions';
 import {columns} from '../database/preguntasCuscoDB';
+import {requestExternalStoragePermission} from '../assets/permissions';
 
 export const FileDownloaderScreen = () => {
   const [selectedYear, setSelectedYear] = useState('2023'); // Año seleccionado por defecto
   const handleSaveDataXlSXNumeric = async () => {
     try {
-      const dataCloudFirestore = await handleGetData();
-      const dataSurvey = dataCloudFirestore.filter(
-        data => data.dataType === 'numeric' && data.year === selectedYear,
-      );
-      const valuesArray = [];
-      dataSurvey.forEach((data, index) => {
-        const values = Object.values(data.results);
-        values.unshift(index + 1);
-        valuesArray.push(values);
-      });
+      const permissions = await requestExternalStoragePermission();
+      if (permissions) {
+        const dataCloudFirestore = await handleGetData();
+        const dataSurvey = dataCloudFirestore.filter(
+          data => data.dataType === 'numeric' && data.year === selectedYear,
+        );
+        const valuesArray = [];
+        dataSurvey.forEach((data, index) => {
+          const values = Object.values(data.results);
+          values.unshift(index + 1);
+          valuesArray.push(values);
+        });
 
-      const filePath = `${RNFS.DownloadDirectoryPath}/cusco_numeric_${selectedYear}.xlsx`;
-      let workbook = null;
+        const filePath = `${RNFS.DownloadDirectoryPath}/cusco_numeric_${selectedYear}.xlsx`;
+        let workbook = null;
 
-      if (await RNFS.exists(filePath)) {
-        await RNFS.unlink(filePath); // Elimina el archivo existente
+        if (await RNFS.exists(filePath)) {
+          await RNFS.unlink(filePath); // Elimina el archivo existente
+        }
+        workbook = XLSX.utils.book_new();
+
+        let sheetName = 'Datos';
+        let worksheet = workbook.Sheets[sheetName];
+
+        if (!worksheet) {
+          worksheet = XLSX.utils.aoa_to_sheet([columns]);
+          XLSX.utils.book_append_sheet(workbook, worksheet, sheetName);
+        }
+
+        XLSX.utils.sheet_add_aoa(worksheet, valuesArray, {
+          header: [],
+          skipHeader: true,
+          origin: -1,
+        });
+
+        const newExcelData = XLSX.write(workbook, {
+          bookType: 'xlsx',
+          type: 'base64',
+        });
+
+        await RNFS.writeFile(filePath, newExcelData, 'base64');
+
+        handleSaveDataXlSXString();
+      }else{
+        Alert.alert('Permisos denegados', 'Permite a la aplicación acceder a tus archivos')
       }
-      workbook = XLSX.utils.book_new();
-
-      let sheetName = 'Datos';
-      let worksheet = workbook.Sheets[sheetName];
-
-      if (!worksheet) {
-        worksheet = XLSX.utils.aoa_to_sheet([columns]);
-        XLSX.utils.book_append_sheet(workbook, worksheet, sheetName);
-      }
-
-      XLSX.utils.sheet_add_aoa(worksheet, valuesArray, {
-        header: [],
-        skipHeader: true,
-        origin: -1,
-      });
-
-      const newExcelData = XLSX.write(workbook, {
-        bookType: 'xlsx',
-        type: 'base64',
-      });
-
-      await RNFS.writeFile(filePath, newExcelData, 'base64');
-
-      handleSaveDataXlSXString();
     } catch (error) {
       console.error('Error al guardar datos en Excel:', error);
     }
